Pay Split Manager, charges and pay splits were being shown with zero amounts sometimes. Backported to: 15.2.22.0;15.3.16.0 Reviewed by: Allen - Note: Adding/subtracting doubles cause accuracy errors in the result.  The changes made to the PaySplitEdit window (changing data types from double to decimal) are likely temporary.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/opendental15.3@12929 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Forms/FormAmountEdit.cs
@@ -8,7 +8,7 @@ using OpenDentBusiness;
 
 namespace OpenDental {
 	public partial class FormAmountEdit:Form {
-		public double Amount;
+		public decimal Amount;
 		private string _text;
 
 		public FormAmountEdit(string text) {
@@ -19,13 +19,13 @@ namespace OpenDental {
 
 		private void FormAmountEdit_Load(object sender,EventArgs e) {
 			labelText.Text=_text;
-			textAmount.Text=POut.Double(Amount);
+			textAmount.Text=POut.Decimal(Amount);
 			textAmount.SelectionStart=0;
 			textAmount.SelectionLength=textAmount.Text.Length;
 		}
 
 		private void butOK_Click(object sender,EventArgs e) {
-			Amount=PIn.Double(textAmount.Text);
+			Amount=PIn.Decimal(textAmount.Text);
 			DialogResult=DialogResult.OK;
 		}
